* wired ApiDoc selection view with ApiDocItem selection * introduced item cache for faster re-selection of previously loaded and parsed documentations * several minor improvements (e.g. GUI, tests, etc)

+++ apidocs-desktop/src/main/flex/org/asforge/apidocs/desktop/model/OptionsModel.as
@@ -39,8 +39,10 @@ package org.asforge.apidocs.desktop.model {
 
         public function deleteApiDoc(apiDoc:ApiDoc):void {
             var itemIndex:int = list.getItemIndex(apiDoc);
-            list.removeItemAt(itemIndex)
-            apiDocService.deleteApiDoc(apiDoc);
+            list.removeItemAt(itemIndex);
+            if (apiDoc.id > 0) {
+                apiDocService.deleteApiDoc(apiDoc);
+            }
         }
 
     }
